// src/generated/models/Cgmp_tasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cgmp_taskscgmp_priority = {
  100000000: 'High',
  100000001: 'Medium',
  100000002: 'Low'
} as const;
export type Cgmp_taskscgmp_priority = keyof typeof Cgmp_taskscgmp_priority;
export const Cgmp_taskscgmp_type = {
  100000000: 'UpdateUATUsers',
  100000001: 'ReviewChange',
  100000002: 'AddPIRNotes',
  100000003: 'CloseChangeWindow',
  100000004: 'ApproveRelease',
  100000005: 'Escalate',
  100000006: 'Other'
} as const;
export type Cgmp_taskscgmp_type = keyof typeof Cgmp_taskscgmp_type;
export const Cgmp_tasksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cgmp_tasksstatecode = keyof typeof Cgmp_tasksstatecode;
export const Cgmp_tasksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cgmp_tasksstatuscode = keyof typeof Cgmp_tasksstatuscode;

export interface Cgmp_tasksBase {
  cgmp_assignedto?: string;
  cgmp_completedat?: string;
  cgmp_duedate?: string;
  cgmp_iscompleted?: boolean;
  cgmp_priority: Cgmp_taskscgmp_priority;
  cgmp_relatedchangeid?: string;
  cgmp_relatedprojectid?: string;
  cgmp_taskid: string;
  cgmp_title: string;
  cgmp_type: Cgmp_taskscgmp_type;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cgmp_tasksstatecode;
  statuscode?: Cgmp_tasksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cgmp_tasks extends Cgmp_tasksBase {
  cgmp_iscompletedname?: string;
  cgmp_priorityname?: string;
  cgmp_typename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}