// src/generated/models/Cgmp_userprofilesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cgmp_userprofilescgmp_notificationpreference = {
  100000000: 'Email',
  100000001: 'Teams',
  100000002: 'Both'
} as const;
export type Cgmp_userprofilescgmp_notificationpreference = keyof typeof Cgmp_userprofilescgmp_notificationpreference;
export const Cgmp_userprofilescgmp_role = {
  100000000: 'Admin',
  100000001: 'PMO',
  100000002: 'LocationITOpsPOC',
  100000003: 'ISM',
  100000004: 'GIICC',
  100000005: 'Observer',
  100000006: 'ISMDeputy',
  100000007: 'DeptAdmin'
} as const;
export type Cgmp_userprofilescgmp_role = keyof typeof Cgmp_userprofilescgmp_role;
export const Cgmp_userprofilesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cgmp_userprofilesstatecode = keyof typeof Cgmp_userprofilesstatecode;
export const Cgmp_userprofilesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cgmp_userprofilesstatuscode = keyof typeof Cgmp_userprofilesstatuscode;

export interface Cgmp_userprofilesBase {
  cgmp_assignedlocations?: string;
  cgmp_assignedprojectids?: string;
  cgmp_isactive?: boolean;
  cgmp_lastseenat?: string;
  cgmp_notificationcategories?: string;
  cgmp_notificationpreference?: Cgmp_userprofilescgmp_notificationpreference;
  cgmp_powerbiurl?: string;
  cgmp_quiethoursend?: number;
  cgmp_quiethoursstart?: number;
  cgmp_role: Cgmp_userprofilescgmp_role;
  cgmp_userprincipalname: string;
  cgmp_userprofileid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cgmp_userprofilesstatecode;
  statuscode?: Cgmp_userprofilesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cgmp_userprofiles extends Cgmp_userprofilesBase {
  cgmp_notificationpreferencename?: string;
  cgmp_rolename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}